requirements.txt
Flask
google-cloud-vision
google-auth
openai
python-dotenv
